Handle timediff from 1h30m as 2 hours.
There was a bug with 1h and 30mins given 1 hours back. This should be 2 hours.
I added a simple check to see if the returning hour was 1, if so, i return 2.
This will now give the indented number of hours.

Changed the test to check, if "2 hours ago" is returned with 90mins and 110mins
instead of just checking for "hours ago".

Issue #79