Handle 61 days ago as 2 months
We have a fun problem with the summer months, where
July and August has 31 days each. This resulted in
the script returning "1 months ago" instead of 2.
The DateTime diff() was not incorrect though, as
it was only 1 month and 31 days ago, but for our
purpose, i have changed the script to round up from
1 to 2, when we are about 59days ago.

Handles #95